+++ src/TurtlePath.Domain/Identifier/CIdProfile.cs
@@ -1,0 +1,11 @@
+namespace TurtlePath.Domain.Identifier
+{
+    /// <summary>
+    /// Base class for CId definition profiles.
+    /// </summary>
+    public abstract class CIdProfile : ICIdProfile
+    {
+        /// <inheritdoc/>
+        public abstract void Configure(CIdProfileBuilder builder);
+    }
+}